requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-celery-beat==2.2.1
django-filter==2.4.0
django-rest-framework==0.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-spectacular==0.17.3
flake8==3.9.2
idna==3.2
inflection==0.5.1
jsonschema==3.2.0
kombu==5.1.0
mccabe==0.6.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyrsistent==0.18.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
